+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [4.0.0-alpha.7] - 2025-11-19
+
+### Added
+
+- `hasErrors`, `assertLinterFixCommand`, `assertLinterSuggestions`,
+  `isLinterFixCommand`, `isLinterSuggestions` helpers.
+
+[4.0.0-alpha.7]: https://github.com/AdguardTeam/AGLint/compare/v4.0.0-alpha.6...v4.0.0-alpha.7
+
 ## [4.0.0-alpha.6] - 2025-11-19
 
 ### Added